requirements.txt
certifi==2021.5.30
chardet==4.0.0
greenlet==1.1.0
idna==2.10
piexif==1.1.3
Pillow==8.2.0
plum-py==0.3.1
pyTelegramBotAPI==3.7.9
python-dotenv==0.17.1
requests==2.25.1
SQLAlchemy==1.4.18
urllib3==1.26.5
